Factor out common csv tools
